requirements.txt
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocs-minify-plugin>=0.8.0
pymdown-extensions>=10.7.0
pyyaml>=6.0  # Necesario para el hook draft_filter.py
